[FIX] 모든 컴포넌트가 전역으로 사용되는 api 패키지의 특성에 맞추어 api.global 패키지를 api.common 으로 변경
