+++ src/main/kotlin/com/iplease/lib/bento/api/common/listener/Listener.kt
@@ -1,4 +1,4 @@
-package com.iplease.lib.bento.api.global.listener
+package com.iplease.lib.bento.api.common.listener
 
 interface Listener<T: Any> {
     fun subscribe(data: T)
