+++ src/main/kotlin/com/iplease/lib/bento/api/common/message/MessageFactory.kt
@@ -1,0 +1,8 @@
+package com.iplease.lib.bento.api.common.message
+
+import com.iplease.lib.bento.api.common.message.payload.Payload
+import com.iplease.lib.bento.api.common.message.routingkey.RoutingKey
+
+interface MessageFactory {
+    fun create(second: RoutingKey<*>, first: Payload<*>): Message
+}
