+++ src/main/kotlin/com/iplease/lib/bento/api/common/message/payload/Payload.kt
@@ -1,0 +1,5 @@
+package com.iplease.lib.bento.api.common.message.payload
+
+interface Payload<T> {
+    fun getValue(): T
+}
