+++ src/main/kotlin/com/iplease/lib/bento/api/common/message/payload/PayloadFactory.kt
@@ -1,4 +1,4 @@
-package com.iplease.lib.bento.api.global.message.payload
+package com.iplease.lib.bento.api.common.message.payload
 
 interface PayloadFactory<T> {
     fun create(value: T): Payload<T>
