+++ src/main/kotlin/com/iplease/lib/bento/api/common/message/routingkey/RoutingKey.kt
@@ -1,0 +1,5 @@
+package com.iplease.lib.bento.api.common.message.routingkey
+
+interface RoutingKey<T> {
+    fun getValue(): T
+}
