+++ src/main/kotlin/com/iplease/lib/bento/api/common/message/routingkey/RoutingKeyFactory.kt
@@ -1,4 +1,4 @@
-package com.iplease.lib.bento.api.global.message.routingkey
+package com.iplease.lib.bento.api.common.message.routingkey
 
 interface RoutingKeyFactory<T> {
     fun create(value: T): RoutingKey<T>
