+++ src/main/kotlin/com/iplease/lib/bento/api/common/message/routingkey/RoutingKeyFormatter.kt
@@ -1,4 +1,4 @@
-package com.iplease.lib.bento.api.global.message.routingkey
+package com.iplease.lib.bento.api.common.message.routingkey
 
 interface RoutingKeyFormatter<T> {
     fun format(data: T): RoutingKey<*>
