+++ src/main/kotlin/com/iplease/lib/bento/api/common/publish/MessagePublishService.kt
@@ -1,0 +1,7 @@
+package com.iplease.lib.bento.api.common.publish
+
+import com.iplease.lib.bento.api.common.message.Message
+
+interface MessagePublishService {
+    fun publish(message: Message)
+}
