+++ src/main/kotlin/com/iplease/lib/bento/api/common/publish/TypedMessagePublishService.kt
@@ -1,4 +1,4 @@
-package com.iplease.lib.bento.api.global.publish
+package com.iplease.lib.bento.api.common.publish
 
 interface TypedMessagePublishService<K, V>: MessagePublishService {
     fun publish(routingKey: K, data: V)
