+++ src/main/kotlin/com/iplease/lib/bento/api/common/subscribe/MessageSubscribeService.kt
@@ -1,0 +1,7 @@
+package com.iplease.lib.bento.api.common.subscribe
+
+import com.iplease.lib.bento.api.common.message.Message
+
+interface MessageSubscribeService {
+    fun subscribe(message: Message)
+}
