--- src/main/kotlin/com/iplease/lib/bento/api/global/message/MessageFactory.kt
@@ -1,8 +1,0 @@
-package com.iplease.lib.bento.api.global.message
-
-import com.iplease.lib.bento.api.global.message.payload.Payload
-import com.iplease.lib.bento.api.global.message.routingkey.RoutingKey
-
-interface MessageFactory {
-    fun create(second: RoutingKey<*>, first: Payload<*>): Message
-}
