--- src/main/kotlin/com/iplease/lib/bento/api/global/message/payload/Payload.kt
@@ -1,5 +1,0 @@
-package com.iplease.lib.bento.api.global.message.payload
-
-interface Payload<T> {
-    fun getValue(): T
-}
