--- src/main/kotlin/com/iplease/lib/bento/api/global/message/routingkey/RoutingKey.kt
@@ -1,5 +1,0 @@
-package com.iplease.lib.bento.api.global.message.routingkey
-
-interface RoutingKey<T> {
-    fun getValue(): T
-}
