--- src/main/kotlin/com/iplease/lib/bento/api/global/publish/MessagePublishService.kt
@@ -1,7 +1,0 @@
-package com.iplease.lib.bento.api.global.publish
-
-import com.iplease.lib.bento.api.global.message.Message
-
-interface MessagePublishService {
-    fun publish(message: Message)
-}
