+++ src/main/kotlin/com/iplease/lib/bento/impl/global/subscribe/RouteMessageSubscribeService.kt
@@ -1,8 +1,8 @@
 package com.iplease.lib.bento.impl.global.subscribe
 
-import com.iplease.lib.bento.api.global.listener.MessageListener
-import com.iplease.lib.bento.api.global.message.Message
-import com.iplease.lib.bento.api.global.subscribe.MessageSubscribeService
+import com.iplease.lib.bento.api.common.listener.MessageListener
+import com.iplease.lib.bento.api.common.message.Message
+import com.iplease.lib.bento.api.common.subscribe.MessageSubscribeService
 
 abstract class RouteMessageSubscribeService: MessageSubscribeService {
     final override fun subscribe(message: Message) { getListeners().forEach { it.subscribe(message) } }
